出力フォーマットの判定場所を変更
- /service/action.json/id/1 のような形に対応

なぜかテストが通らなくなった（runTestsInSeparateProcessesがおかしい？）

+++ tests/ellsif/WelCMS/UserRepositoryTest.php
@@ -22,7 +22,9 @@ class UserRepositoryTest extends \PHPUnit\Framework\TestCase
 
     public static function tearDownAfterClass()
     {
-        unlink(dirname(__FILE__, 3) . '/data/UserRepositoryTest.sqlite');
+        if (file_exists(dirname(__FILE__, 3) . '/data/UserRepositoryTest.sqlite')) {
+            unlink(dirname(__FILE__, 3) . '/data/UserRepositoryTest.sqlite');
+        }
     }
 
     protected function setUp()
